==> notebooks/00_parse_uniprot_sprot_xml.ipynb <==
%load_ext autoreload
%autoreload 2
import polars as pl
from parse_uniprot_xml_taxonomy import process_uniprot_file

## Parse UniProtKB/SwissProt (reviewed, high-quality annotations)

Download [Reviwed (Swiss-Prot) xml](https://ftp.uniprot.org/pub/databases/uniprot/current_release/knowledgebase/complete/uniprot_sprot.xml.gz) from https://www.uniprot.org/help/downloads and save to `../data/`


Then process with the below function. This takes ~2-3 minutes to run

input_file = "../data/uniprot_sprot.xml.gz"
output_file = "../data/uniprot_sprot_taxonomy_summary.parquet"
process_uniprot_file(input_file, output_file)

uniprot_sprot = pl.read_parquet(output_file)
uniprot_sprot.head()

### Total PDB structures

uniprot_sprot["pdb_structures_count"].sum()

### Human has the most, 8511 sequences -> What percentage is this?

100 * 8511 / uniprot_sprot["pdb_structures_count"].sum()
