==> notebooks/01_parse_uniprot_trembl_xml.ipynb <==
## Parse UniProtKB/TREMBL (unreviewed, computationally-derived annotations)

We created a TSV of UniProt entries using the [uniprot](https://github.com/heuermh/dishevelled-bio/tree/master/protein/src/main/java/org/dishevelled/bio/protein/uniprot) tool from [heuermh/dishevelled-bio](https://github.com/heuermh/dishevelled-bio). This created a 40 GB TSV file from the 219 GB gzipped XML file `uniprot_trembl.xml.gz` downloaded directly from [UniProt](https://www.uniprot.org/help/downloads).

This took ~5-6 hours to run.

! uniprot-1.0-SNAPSHOT/bin/uniprot -i ../data/uniprot_trembl.xml.gz > ../data/uniprot_trembl_entries.tsv
