==> notebooks/02_add_kingdom_domains_to_sprot.ipynb <==
import numpy as np
import pandas as pd
import plotly.graph_objects as go
import polars as pl
import seaborn as sns
from ipywidgets import widgets

uniprot_sprot = pd.read_parquet("../data/uniprot_sprot_taxonomy_summary.parquet")
print(uniprot_sprot.shape)


def merge_microbes(species_type):
    if species_type in {"Bacteria", "Archaea", "Viruses"}:
        return "Microbial"
    else:
        return species_type.title()


def assign_superdomain(species_type):
    if species_type == "Viruses":
        return "Non-cellular Life"
    else:
        return "Cellular Life"


def assign_domain(species_type):
    if species_type in {"Bacteria", "Archaea", "Viruses"}:
        return species_type
    else:
        return "Eukaryota"


def assign_kingdom(species_type):
    if species_type in {"Bacteria", "Archaea"}:
        return "Monera"
    else:
        return species_type


uniprot_sprot["type_merge_microbes"] = uniprot_sprot["type"].apply(merge_microbes)
uniprot_sprot["superdomain"] = uniprot_sprot["type"].apply(assign_superdomain)
uniprot_sprot["domain"] = uniprot_sprot["type"].apply(assign_domain)
uniprot_sprot["kingdom"] = uniprot_sprot["type"].apply(assign_kingdom)
uniprot_sprot.head()

### What percentage of the structures are from human?

uniprot_sprot.pdb_structures_count.sum()

8511 / uniprot_sprot.pdb_structures_count.sum()

uniprot_sprot.to_parquet("../data/uniprot_sprot_taxonomy_summary_with_domains.parquet")

uniprot_sprot.sort_values("reviewed_count", ascending=False).head(10)

uniprot_sprot.sort_values("pdb_structures_count", ascending=False).head(10)

uniprot_sprot.groupby(["superdomain", "domain", "kingdom"]).size()

# uniprot_sprot.sort_values("n_proteins", ascending=False).head()

uniprot_sprot_summarized_per_organism = uniprot_sprot.groupby("organism").agg(
    {
        "reviewed_count": "sum",
        "organism": "first",
        "organism_id": "first",
        "lineage": "first",
        "type": "first",
        "type_merge_microbes": "first",
        "pdb_structures_count": "sum",
    }
)
print(uniprot_sprot_summarized_per_organism.shape)
uniprot_sprot_summarized_per_organism.sort_values("reviewed_count", ascending=False)

uniprot_sprot_n_proteins_per_organism = (
    uniprot_sprot.groupby("organism").reviewed_count.sum().sort_values(ascending=False)
)
uniprot_sprot_n_proteins_per_organism.head(10)

uniprot_sprot.type_merge_microbes.value_counts()

uniprot_sprot_n_proteins = uniprot_sprot.groupby(
    "type_merge_microbes"
).reviewed_count.sum()
uniprot_sprot_n_proteins.name = "n_proteins_sprot"
uniprot_sprot_n_proteins

### Sanity check that this is equal to the expected number of UniProt/SwissProt Proteins for Release  2025_01

assert uniprot_sprot_n_proteins.sum() == 572970
