==> notebooks/06_combine_uniprot_swissprot_trembl_total_pdb_structures.ipynb <==
import polars as pl

unprot_sprot = pl.scan_parquet(
    "../data/uniprot_sprot_taxonomy_summary_with_domains.parquet"
)
unprot_sprot.head().collect()

unprot_trembl = pl.scan_parquet("../data/uniprot_trembl_taxonomy_summary.parquet")

unprot_trembl.head().collect()

"organism	organism_id	lineage	type type_merge_microbes	superdomain	domain	kingdom".split()

%%time

uniprot_combined = (
    pl.concat([unprot_sprot, unprot_trembl.select(unprot_sprot.columns)])
    .group_by(
        [
            "organism",
            "organism_id",
            "lineage",
            "type",
            "type_merge_microbes",
            "superdomain",
            "domain",
            "kingdom",
        ]
    )
    .agg(
        pl.col("pdb_structures_count").sum(),
        pl.col("reviewed_count").sum(),
        pl.col("unreviewed_count").sum(),
    )
    .sort("pdb_structures_count", descending=True)
    .collect()
)
uniprot_combined

### How many total proteins with a PDB structure, across both reviewed and unreviewed?

uniprot_combined["pdb_structures_count"].sum()

uniprot_combined.write_parquet(
    "../data/uniprot_combined_sprot_trembl_taxonomy_summary.parquet"
)

uniprot_combined.select(pl.col(["reviewed_count", "unreviewed_count"]).sum())

### How many from human? -> Human has 8511 proteins with PDB structures

100 * 8511 / uniprot_combined["pdb_structures_count"].sum()

## What percent of all proteins does uniprot represent?

total_uniprot = 252633201 + 572970
f"{total_uniprot:,}"

n_predicted_proteins = 14e12
f"{n_predicted_proteins:,}"

total_uniprot / n_predicted_proteins

### Show as a percent

f"{100*total_uniprot /n_predicted_proteins:.5f}"
