==> notebooks/06_convert_uniprot_trembl_to_parquet.ipynb <==
import gzip

import polars as pl

## Convert tsv to parquet

We created a TSV of UniProt entries using the [uniprot](https://github.com/heuermh/dishevelled-bio/tree/master/protein/src/main/java/org/dishevelled/bio/protein/uniprot) tool from [heuermh/dishevelled-bio](https://github.com/heuermh/dishevelled-bio). This created a 40 GB TSV file from the 219 GB gzipped XML file `uniprot_trembl.xml.gz` downloaded directly from [UniProt](https://www.uniprot.org/help/downloads).

```
uniprot-1.0-SNAPSHOT/bin/uniprot -i uniprot_trembl.xml.gz > uniprot_trembl_entries.tsv
```

First, let's convert this to parquet file so it is more compact (small enough for this github repo!) and better for streaming.

%%time

# with gzip.open("uniprot_trembl_entries.tsv") as f:
df = pl.scan_csv(
    "../data/uniprot_trembl_entries.tsv",
    separator="\t",
    new_columns=[
        "organism",
        "organism_id",
        "lineage",
        "type",
        "reviewed",
        "unreviewed",
        "has_structure",
    ],
    has_header=False,
)
df.sink_parquet("../data/uniprot_trembl_entries.parquet")

df.head().collect()

# Read in parquet

uniprot_trembl = pl.scan_parquet("../data/uniprot_trembl_entries.parquet")
uniprot_trembl

## Assign kingdoms, domains, summarize per organism


def merge_microbes(species_type):
    if species_type in {"Bacteria", "Archaea", "Viruses"}:
        return "Microbial"
    else:
        return species_type.title()


def assign_superdomain(species_type):
    if species_type == "Viruses":
        return "Non-cellular Life"
    else:
        return "Cellular Life"


def assign_domain(species_type):
    if species_type in {"Bacteria", "Archaea", "Viruses"}:
        return species_type
    else:
        return "Eukaryota"


def assign_kingdom(species_type):
    if species_type in {"Bacteria", "Archaea"}:
        return "Monera"
    else:
        return species_type


uniprot_trembl = uniprot_trembl.with_columns(
    pl.col("type")
    .map_elements(merge_microbes, return_dtype=pl.String)
    .alias("type_merge_microbes"),
    pl.col("type")
    .map_elements(assign_superdomain, return_dtype=pl.String)
    .alias("superdomain"),
    pl.col("type").map_elements(assign_domain, return_dtype=pl.String).alias("domain"),
    pl.col("type")
    .map_elements(assign_kingdom, return_dtype=pl.String)
    .alias("kingdom"),
)
uniprot_trembl.head().collect()

%%time


def merge_microbes(species_type):
    if species_type in {"Bacteria", "Archaea", "Viruses"}:
        return "Microbial"
    else:
        return species_type.title()


def assign_superdomain(species_type):
    if species_type == "Viruses":
        return "Non-cellular Life"
    else:
        return "Cellular Life"


def assign_domain(species_type):
    if species_type in {"Bacteria", "Archaea", "Viruses"}:
        return species_type
    else:
        return "Eukaryota"


def assign_kingdom(species_type):
    if species_type in {"Bacteria", "Archaea"}:
        return "Monera"
    else:
        return species_type


uniprot_trembl_summarized_per_organism = (
    uniprot_trembl.group_by(["organism", "organism_id", "lineage", "type"])
    .agg(
        [
            pl.col("reviewed").sum().alias("reviewed_count"),
            pl.col("unreviewed").sum().alias("unreviewed_count"),
            pl.col("has_structure")
            .sum()
            .alias("pdb_structures_count"),  # Renamed to be explicit
        ]
    )
    .with_columns(
        pl.col("type")
        .map_elements(merge_microbes, return_dtype=pl.String)
        .alias("type_merge_microbes"),
        pl.col("type")
        .map_elements(assign_superdomain, return_dtype=pl.String)
        .alias("superdomain"),
        pl.col("type")
        .map_elements(assign_domain, return_dtype=pl.String)
        .alias("domain"),
        pl.col("type")
        .map_elements(assign_kingdom, return_dtype=pl.String)
        .alias("kingdom"),
    )
    .collect()
)
print(uniprot_trembl_summarized_per_organism.shape)
uniprot_trembl_summarized_per_organism.sort("reviewed_count", descending=True)

### Curious: How many unreviewed have PDB structures?

uniprot_trembl_summarized_per_organism["pdb_structures_count"].sum()

### Sanity check: this should be ~250M sequences

uniprot_trembl_summarized_per_organism["unreviewed_count"].sum()

assert uniprot_trembl_summarized_per_organism["unreviewed_count"].sum() == 252633201

### Write to file!

uniprot_trembl_summarized_per_organism.write_parquet(
    "../data/uniprot_trembl_taxonomy_summary.parquet"
)
